feat: add projections for ParcelAddressWasReplacedBecauseOfMunicipalityMerger

+++ src/ParcelRegistry.Projections.Integration/ParcelLatestItem/ParcelLatestItemProjections.cs
@@ -1,5 +1,8 @@
 ﻿namespace ParcelRegistry.Projections.Integration.ParcelLatestItem
 {
+    using System;
+    using System.Threading;
+    using System.Threading.Tasks;
     using Be.Vlaanderen.Basisregisters.ProjectionHandling.Connector;
     using Be.Vlaanderen.Basisregisters.ProjectionHandling.SqlStreamStore;
     using Converters;
@@ -104,6 +107,18 @@
                         message.Message.CaPaKey), ct);
             });
 
+            When<Envelope<ParcelAddressWasReplacedBecauseOfMunicipalityMerger>>(async (context, message, ct) =>
+            {
+                await RemoveParcelAddress(context, message.Message.ParcelId, message.Message.PreviousAddressPersistentLocalId, ct);
+
+                await context
+                    .ParcelLatestItemAddresses
+                    .AddAsync(new ParcelLatestItemAddress(
+                        message.Message.ParcelId,
+                        message.Message.NewAddressPersistentLocalId,
+                        message.Message.CaPaKey), ct);
+            });
+
             When<Envelope<ParcelAddressWasReplacedBecauseAddressWasReaddressed>>(async (context, message, ct) =>
             {
                 var previousAddress = await context
@@ -142,14 +157,7 @@
             {
                 foreach (var addressPersistentLocalId in message.Message.DetachedAddressPersistentLocalIds)
                 {
-                    var relation = await context
-                        .ParcelLatestItemAddresses
-                        .FindAsync([message.Message.ParcelId, addressPersistentLocalId], ct);
-
-                    if (relation is not null)
-                    {
-                        context.ParcelLatestItemAddresses.Remove(relation);
-                    }
+                    await RemoveParcelAddress(context, message.Message.ParcelId, addressPersistentLocalId, ct);
                 }
 
                 foreach (var addressPersistentLocalId in message.Message.AttachedAddressPersistentLocalIds)
@@ -172,39 +180,39 @@
 
             When<Envelope<ParcelAddressWasDetachedV2>>(async (context, message, ct) =>
             {
-                var latestItemAddress = await context
-                    .ParcelLatestItemAddresses
-                    .FindAsync(new object?[] { message.Message.ParcelId, message.Message.AddressPersistentLocalId }, cancellationToken: ct);
-
-                context.ParcelLatestItemAddresses.Remove(latestItemAddress);
+                await RemoveParcelAddress(context, message.Message.ParcelId, message.Message.AddressPersistentLocalId, ct);
             });
 
             When<Envelope<ParcelAddressWasDetachedBecauseAddressWasRejected>>(async (context, message, ct) =>
             {
-                var latestItemAddress = await context
-                    .ParcelLatestItemAddresses
-                    .FindAsync(new object?[] { message.Message.ParcelId, message.Message.AddressPersistentLocalId }, cancellationToken: ct);
-
-                context.ParcelLatestItemAddresses.Remove(latestItemAddress);
+                await RemoveParcelAddress(context, message.Message.ParcelId, message.Message.AddressPersistentLocalId, ct);
             });
 
             When<Envelope<ParcelAddressWasDetachedBecauseAddressWasRemoved>>(async (context, message, ct) =>
             {
-                var latestItemAddress = await context
-                    .ParcelLatestItemAddresses
-                    .FindAsync(new object?[] { message.Message.ParcelId, message.Message.AddressPersistentLocalId }, cancellationToken: ct);
-
-                context.ParcelLatestItemAddresses.Remove(latestItemAddress);
+                await RemoveParcelAddress(context, message.Message.ParcelId, message.Message.AddressPersistentLocalId, ct);
             });
 
             When<Envelope<ParcelAddressWasDetachedBecauseAddressWasRetired>>(async (context, message, ct) =>
             {
-                var latestItemAddress = await context
-                    .ParcelLatestItemAddresses
-                    .FindAsync(new object?[] { message.Message.ParcelId, message.Message.AddressPersistentLocalId }, cancellationToken: ct);
-
-                context.ParcelLatestItemAddresses.Remove(latestItemAddress);
+                await RemoveParcelAddress(context, message.Message.ParcelId, message.Message.AddressPersistentLocalId, ct);
             });
+        }
+
+        private static async Task RemoveParcelAddress(
+            IntegrationContext context,
+            Guid parcelId,
+            int addressPersistentLocalId,
+            CancellationToken ct)
+        {
+            var latestItemAddress = await context
+                .ParcelLatestItemAddresses
+                .FindAsync(new object?[] { parcelId, addressPersistentLocalId }, cancellationToken: ct);
+
+            if (latestItemAddress is not null)
+            {
+                context.ParcelLatestItemAddresses.Remove(latestItemAddress);
+            }
         }
 
         private static void UpdateVersionTimestamp(ParcelLatestItem parcel, Instant versionTimestamp)
